% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2latex.R
\name{attr2text}
\alias{attr2text}
\title{get attributes as text}
\usage{
attr2text(
  dt,
  attr = c("rows", "weight", "units", "info"),
  units.name = NULL,
  sep = ". ",
  vector = FALSE,
  rm.if.all = FALSE
)
}
\arguments{
\item{dt}{a dtable}

\item{attr}{which attributes to extract}

\item{units.name}{character, name of units}

\item{sep}{separator between attributes for output}

\item{vector}{return vector instead of a single string}

\item{rm.if.all}{exclude info on units if all unique}
}
\description{
extract attributes of dtable to a string (or character vector)
}
