% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp-BL-diff.R
\name{baseline-compare-diff}
\alias{baseline-compare-diff}
\alias{real.diff.bl}
\alias{catg.diff.bl}
\alias{bnry.diff.bl}
\alias{lcat.diff.bl}
\alias{date.diff.bl}
\alias{surv.diff.bl}
\title{default baseline comparer 'diff'}
\usage{
real.diff.bl(x, g, weight = NULL, ...)

catg.diff.bl(x, g, weight = NULL, ...)

bnry.diff.bl(x, g, weight = NULL, ...)

lcat.diff.bl(x, g, weight = NULL, ...)

date.diff.bl(x, g, weight = NULL, ...)

surv.diff.bl(time, event, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
Comparer functions to calculate differences that work together.

real.diff.bl: (weighted) 'real' standardized difference

catg.diff.bl: (weighted) 'catg' standardized difference

bnry.diff.bl: (weighted) standardized difference for 'bnry'

lcat.diff.bl: empty

date.diff.bl: (weighted) standardized difference for 'date' ( = real.diff
applied to x interpreted as an integer)

surv.diff.bl: (weighted) standardized difference for 'surv'
}
