% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp-BL-std.R
\name{baseline-compare-std}
\alias{baseline-compare-std}
\alias{real.std.bl}
\alias{catg.std.bl}
\alias{bnry.std.bl}
\alias{lcat.std.bl}
\alias{date.std.bl}
\alias{surv.std.bl}
\title{default baseline comparer 'std'}
\usage{
real.std.bl(x, g, weight = NULL, ...)

catg.std.bl(x, g, weight = NULL, ...)

bnry.std.bl(x, g, weight = NULL, ...)

lcat.std.bl(x, g, weight = NULL, ...)

date.std.bl(x, g, weight = NULL, ...)

surv.std.bl(time, event, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
Comparer functions to calculate standardized differences that work together.

real.std.bl: (weighted) 'real' standardized difference

catg.std.bl: (weighted) 'catg' standardized difference

bnry.std.bl: (weighted) standardized difference for 'bnry'

lcat.std.bl: empty

date.std.bl: (weighted) standardized difference for 'date' ( = real.std
applied to x interpreted as an integer)

surv.std.bl: (weighted) standardized difference for 'surv'
}
