% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-baseline.R
\name{baseline-standard}
\alias{baseline-standard}
\alias{real.bl}
\alias{real2.bl}
\alias{bnry.bl}
\alias{catg.bl}
\alias{lcat.bl}
\alias{date.bl}
\alias{surv.bl}
\title{standard baseline functions}
\usage{
real.bl(x, weight = NULL, ...)

real2.bl(x, weight = NULL, ...)

bnry.bl(x, weight = NULL, ...)

catg.bl(x, weight = NULL, ...)

lcat.bl(x, ...)

date.bl(x, ...)

surv.bl(time, event, weight = NULL, time.unit = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}

\item{...}{arguments passed along}
}
\description{
Description functions that work together
}
\details{
real.bl: (weighted) median and interquartile range

real2.bl: (weighted) mean and standard deviation

bnry.bl: (weighted) count and percent of
non-reference value as single string

catg.bl: (weighted) count and proportion of
categorical levels as single string

lcat.bl: number of unique non-missing values in x

date.bl: min/max

surv.bl: events; rate
}
