% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{baseline}
\alias{baseline}
\alias{bl}
\title{create a baseline table}
\usage{
baseline(data, theme = 0, ...)

bl(data, theme = 0, ...)
}
\arguments{
\item{data}{data.frame of data}

\item{theme}{list of 1 or 2, args passed to \code{bl_theme}}

\item{...}{arguments passed to \code{dable}}
}
\description{
This is a wrapper for bl_theme and dable (with type = 'baseline').
}
