% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc-BL-1.R
\name{bl_theme}
\alias{bl_theme}
\title{set baseline functions}
\usage{
bl_theme(desc = 0, comp = "std", test = NULL)
}
\arguments{
\item{desc}{the 'theme' for descriptives (0,1,2)}

\item{comp}{the 'theme' for comparisons (std, diff)}

\item{test}{the 'theme' for testers (not yet implemented)}
}
\description{
Quickly set baseline functions to some predefined set of functions.
}
