% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2latex.R
\name{blatex}
\alias{blatex}
\alias{blatex0}
\alias{blatex1}
\alias{blatex2}
\title{LaTeX code for themed baseline tables}
\usage{
blatex(bl, ..., .blatex.theme.strict = 1)

blatex0(
  bl,
  format = TRUE,
  count = "rows",
  kill = "term",
  grey = "term",
  file = "",
  where = "hbt",
  row.group = NULL,
  insert.bottom = TRUE,
  ...
)

blatex1(
  bl,
  format = TRUE,
  count = "rows",
  kill = "term",
  grey = "term",
  file = "",
  where = "hbt",
  row.group = NULL,
  insert.bottom = TRUE,
  ...
)

blatex2(bl, ...)
}
\arguments{
\item{bl}{object created with baseline}

\item{...}{arguments passed}

\item{.blatex.theme.strict}{numeric (mostly internal); how upset this
function will be if the 'theme' attribute is missing from the
bl-argument before passing the problem along to blatex_default: 0 = not
all all, 1 = somewhat, cause a warning, 2 = very, cause an error}
}
\description{
LaTeX code for baseline table created with \code{baseline}
}
\details{
blatex0: alias for blatex_default

blatex1: alias for blatex_default

blatex2: code for baseline theme 2 (experimental)
}
