% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-all}
\alias{comp-all}
\alias{empty.std}
\alias{empty.std.bl}
\alias{empty.diff}
\alias{empty.diff.bl}
\title{general comparers}
\usage{
empty.std(...)

empty.std.bl(...)

empty.diff(...)

empty.diff.bl(...)
}
\arguments{
\item{...}{arguments passed}

\item{x}{input vector}

\item{weight}{case weight}

\item{g}{grouping variable}
}
\description{
functions to compare any variables

empty.std: empty standardized difference calculated

empty.std.bl: empty standardized difference calculated (for baseline tables)

empty.diff: empty standardized difference calculated

empty.diff.bl: empty standardized difference calculated (for baseline tables)
}
