% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-catg}
\alias{comp-catg}
\alias{catg.std}
\alias{catg.std.bl}
\alias{catg.diff}
\alias{catg.diff.bl}
\title{'catg' comparers}
\usage{
catg.std(x, g, weight = NULL, ...)

catg.std.bl(x, g, weight = NULL, ...)

catg.diff(x, g, weight = NULL, ...)

catg.diff.bl(x, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to compare catg variables

catg.std: (weighted) 'catg' standardized difference

catg.std.bl: (weighted) 'catg' standardized difference (for baseline)

catg.diff: (weighted) difference in proportion

catg.diff: (weighted) difference in proportion for baseline table
}
