% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-real}
\alias{comp-real}
\alias{real.std}
\alias{real.std.bl}
\alias{real.diff}
\alias{real.diff.bl}
\title{'real' comparers}
\usage{
real.std(x, g, weight = NULL, ...)

real.std.bl(x, g, weight = NULL, ...)

real.diff(x, g, weight = NULL, ...)

real.diff.bl(x, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to compare real variables

real.std: (weighted) 'real' standardized difference

real.std.bl: (weighted) 'real' standardized difference (for
baseline tables)

real.diff: (weighted) difference in mean value

real.diff.bl: (weighted) difference in mean value (for baseline tables)
}
