% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{dable-formatting}
\alias{dable-formatting}
\alias{dafonumb}
\alias{dafonumb1}
\alias{dafotext}
\alias{dafotext1}
\title{dable formatting functions}
\usage{
dafonumb(x, ...)

dafonumb1(
  x,
  digits = NULL,
  scientific = NULL,
  p = NULL,
  p.bound = NULL,
  NAtext = NULL,
  no04int = NULL,
  ...
)

dafotext(x, ...)

dafotext1(x, output = dpget("output"), substitute = dpget("substitute"), ...)
}
\arguments{
\item{x}{input}

\item{...}{arguments passed}

\item{digits}{number of digits presented after formatting? Vector of length
three to give number of digits for small (absolute value smaller than
1), medium (absolute value between 1 and 10) and large (absolute value
larger than 10) numbers. Default: c("small" = 2, "mid" = 2, "large" = 1)}

\item{scientific}{if to use scientific number representation, and what the
cut offs should be, default: list(use = TRUE, low = 1e-8, high = 1e8)}

\item{p}{are numbers p-value? default: FALSE}

\item{p.bound}{cut off for p-value presentation}

\item{NAtext}{how to show NA in tables}

\item{no04int}{logical; should an integer value be printed without zeros?
(XK add details as to why this is even a thing)}

\item{output}{what is the output? e.g. "latex"}

\item{substitute}{the table of substitutions}
}
\description{
What can we say about formatting? (XK add details)
}
\details{
dafonumb: format a numeric/integer vector

dafonumb1: format a single numeric/integer (helper function for dafonumb)

dafotext: format a vector of strings

dafotext1: format a single string (helper function for dafotext)
}
