% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-internal.R
\name{dable-internal}
\alias{dable-internal}
\alias{di.dots}
\alias{di.Variable}
\title{internal helper functions}
\usage{
di.dots(iinfo, strict = FALSE, ...)

di.Variable(x, ...)
}
\arguments{
\item{iinfo}{character, the name of internal info accessible to functions}

\item{strict}{logical, if TRUE return NULL when iinfo not found, else
non-empty string}

\item{...}{arguments passed from within 'dable'}

\item{x}{the input variable}
}
\description{
functions that work properly inside 'dable'
}
\details{
di.dots: extract info from the '...'

di.Variable: create a 'Variable' column
}
