% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy_defaults.R
\name{dable-parameters}
\alias{dable-parameters}
\alias{dpset}
\alias{dpset_defaults}
\alias{dpget}
\alias{dpget_all}
\alias{add2bnry.list}
\title{dable parameters and how to set and get them}
\usage{
dpset(param, value = NULL)

dpset_defaults(overwrite = TRUE, style = NA_character_)

dpget(param, verbose = FALSE, sub = TRUE)

dpget_all()

add2bnry.list(bl)
}
\arguments{
\item{param}{character; name of parameter OR a named list of parameter values}

\item{value}{the value to be set}

\item{overwrite}{logical; overwrite parameter values already set?}

\item{verbose}{logical; get (possibly) helpful messages?}

\item{sub}{logical; search for possible replacement values?}

\item{bl}{list of binary values}
}
\description{
dable parameters and how to set and get them
}
\details{
These are functions to set and retrieve dable parameters. The
parameters in the package are:
\itemize{
\item \emph{bnry.list} the list of values defining the bnry type
\item \emph{real.tol} the threshold for number of unique numeric values to be
considered to be type catg
\item \emph{catg.tol} the threshold for number of unique categorical values to be
considered to be type lcat
\item \emph{surv.prefix} are surv variables identifiable through a prefix? (else suffix)
\item \emph{surv.affix} what is the affix? default: c("time" = "t.", "event" = "ev.")
\item \emph{id} character vector of id variable names, default: NULL
\item \emph{gtab.defvar.rm} if gtab is specified by a term, should that term be hidden
in output?
\item \emph{weight.defvar.rm} if weight is specified by a term (must be, currently),
should that term be hidden in output?
\item \emph{unit.id.rm} should the unit (first element of 'id') id be hidden in output?
\item \emph{vtab.group.name} default grouping name (for all but surv type) if not given
\item \emph{stab.group.name} default grouping name for surv type if not given
\item \emph{gtab.group.name} default gtab 'all' name if not given
\item \emph{comp.header} header for comparison column i baseline tables
\item \emph{test.header} header for test column i baseline tables
\item \emph{units.name} how to refer to the units in text, default: subjects
\item \emph{grey.first} if using alternating grey in output table, begin with a grey line?
\item \emph{digits} number of digits presented after formatting? Vector of length
three to give number of digits for small (absolute value smaller than 1),
medium (absolute value between 1 and 10) and large (absolute value larger
than 10) numbers. Default: c("small" = 2, "mid" = 2, "large" = 1)
\item \emph{scientific} if to use scientific number representation, and what the cut
offs should be, default: list(use = TRUE, low = 1e-8, high = 1e8)
\item \emph{p} for formatting, are numbers p-value? default: FALSE (but this is
generally automatically identified by formatting functions)
\item \emph{p.bound} cut off for p-value presentation
\item \emph{NAtext} how to show NA in tables
\item \emph{NAalias} how to refer to missing
\item \emph{sep} how to separate the concatenation of variable name and its level
(when applicable)
\item \emph{output} (NOT IN USE) the type of output
\item \emph{indent} what string to use as indentation, default "    ", but
recommended to set to "\\quad" if LaTeX output is wanted
\item \emph{percent} (NOT IN USE) symbol to use for percent
\item \emph{date.format} the format used for dates, default "\%y\%m\%d"
\item \emph{type.part} where type is any type (real,catg,...) and part is the table
part (desc, comp, test); these set the default such function for simple
descriptive tables
\item \emph{type.part.bl} where type is any type (real,catg,...) and part is the table
part (desc, comp, test); these set the default such function for baseline tables
}

dpset: set a dable package parameter

dpset_defaults: set default parameter values

dpget: get package parameter

dpget_all: get all package parameters

add2bnry.list: add a value to the 'bnry.list'
}
