% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2flextable.R
\name{dable2flextable}
\alias{dable2flextable}
\alias{dextable}
\alias{blextable}
\title{functions for flextable output}
\usage{
dextable(
  dt,
  format = TRUE,
  count = "rows",
  kill = NULL,
  t2l = TRUE,
  grey = "term",
  row.group = NULL,
  insert.bottom = TRUE,
  fontsize = c(11, 11, 9)
)

blextable(
  dt,
  format = TRUE,
  count = "rows",
  kill = "term",
  grey = "term",
  row.group = NULL,
  fontsize = c(11, 11, 9),
  insert.bottom = TRUE
)
}
\arguments{
\item{dt}{object from dable}

\item{format}{logical; format the variables in dt?}

\item{count}{character; what entity is presentet in row header 'rows',
'units' or 'weight'}

\item{kill}{character vector; columns to remove for presentation}

\item{t2l}{logical (terms-to-label); replace terms with labels, if applicable}

\item{grey}{character; which rows to make grey-ish in presentation}

\item{row.group}{logical or NULL; use the grouping in guide? If NULL this
will be algorithmically decided}

\item{insert.bottom}{character or logical; text to be placed under the
table. If TRUE, attr2text will be used to create this text}

\item{fontsize}{numerical vector of length 3 for fontsize of header, body
and footnotes}
}
\description{
functions for flextable output
}
\details{
dextable: turn simple descriptive tables into flextables

blextable: turn baseline tables into flextables
}
