% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2latex.R
\name{dable_fnote}
\alias{dable_fnote}
\title{create (latex) footnotes from table variables}
\usage{
dable_fnote(dt, info, fn.var, info.attr = "info", format = FALSE)
}
\arguments{
\item{dt}{a dable}

\item{info}{variable with the footnote info}

\item{fn.var}{variable to get footnotes}

\item{info.attr}{name of attribute to store info in}

\item{format}{format the dable?}
}
\description{
make \code{unique(variable)} a part of attribute 'info' and
place a footnote on another variable
}
