% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2latex.R
\name{dable_format}
\alias{dable_format}
\title{formatting}
\usage{
dable_format(dt, output = dpget("output"))
}
\arguments{
\item{dt}{a dable or such}

\item{output}{output format}
}
\description{
formatting of dables and data.frames
}
