% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2latex.R
\name{dable_prune}
\alias{dable_prune}
\title{prune dable}
\usage{
dable_prune(
  x,
  rm = NULL,
  info = FALSE,
  info.attr = "info",
  info.unique = TRUE,
  split.unique = TRUE
)
}
\arguments{
\item{x}{object}

\item{rm}{index or variable name to remove}

\item{info}{store the discarded information in attributes?}

\item{info.attr}{name of attribute to store discarded info (if \code{info = TRUE})}

\item{info.unique}{store only unique info (if \code{info = TRUE})}

\item{split.unique}{if \code{unique.info = TRUE}, also split into individual
sentences before determining uniqueness?}
}
\description{
prune dable
}
\details{
remove columns by name or index
}
