% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2latex.R
\name{datex.bl}
\alias{datex.bl}
\title{LaTeX code for a dable}
\usage{
datex.bl(
  dt,
  format = TRUE,
  count = "rows",
  kill = "term",
  grey = "term",
  file = "",
  where = "hbt",
  row.group = NULL,
  insert.bottom = TRUE,
  ...
)
}
\arguments{
\item{dt}{object created by dable}

\item{format}{logical; format the variables in dt?}

\item{count}{character; what entity is presentet in row header 'rows',
'units' or 'weight'}

\item{kill}{character vector; columns to remove for presentation}

\item{grey}{character; which rows to make grey-ish in presentation}

\item{file}{character; argument for Hmisc::latex}

\item{where}{character; argument for Hmisc::latex}

\item{row.group}{logical or NULL; use the grouping in guide? If NULL this
will be algorithmically decided}

\item{insert.bottom}{character or logical; text to be placed under the
table. If TRUE, attr2text will be used to create this text}

\item{...}{arguments passed to Hmisc::latex}
}
\description{
A wrapper for Hmisc::latex trying to create nice LaTeX code to present a
dable
}
