% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2latex.R
\name{datex}
\alias{datex}
\title{LaTeX code for a dable}
\usage{
datex(
  dt,
  format = TRUE,
  size = "rows",
  kill = "term",
  grey = "term",
  bl.method = "standard",
  file = "",
  where = "hbt",
  rowname = "Variable",
  row.group = NULL,
  rgroup = NULL,
  insert.bottom = TRUE,
  ...
)
}
\arguments{
\item{dt}{object created by dable}

\item{format}{logical; format the variables in dt?}

\item{size}{character; what entity is presentet in row header 'rows',
'units' or 'weight'}

\item{kill}{character vector; columns to remove for presentation}

\item{grey}{character; which rows to make grey-ish in presentation}

\item{bl.method}{if dable is a 'baseline', what method to use for
presentation? This is a vague feature so far and only a 'standard'
method is implemented.}

\item{file}{character; argument for Hmisc::latex}

\item{where}{character; argument for Hmisc::latex}

\item{rowname}{character; column name to use as rowname in Hmisc::latex}

\item{row.group}{logical or NULL; use the grouping in guide? If NULL this
will be algorithmically decided}

\item{rgroup}{character; argument for Hmisc::latex}

\item{insert.bottom}{character or logical; text to be placed under the
table. If TRUE, attr2text will be used to create this text}

\item{...}{arguments passed to Hmisc::latex}
}
\description{
A wrapper for Hmisc::latex trying to create nice LaTeX code to present a
dable
}
