% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc.R
\name{desc-all}
\alias{desc-all}
\alias{d.n}
\alias{d.length}
\alias{d.sum}
\alias{n.unique}
\alias{n.NA}
\alias{d.mode}
\title{general describers}
\usage{
d.n(x, weight = NULL, ...)

d.length(x, ...)

d.sum(x, weight = NULL, ...)

n.unique(x, ...)

n.NA(x, ...)

d.mode(x, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to describe any variable
}
\details{
d.n: length of non-missing x or sum of weights (if given)

d.length: length of vector

d.sum: (weighted) sum of vector

n.unique: number of unique non-missing values in x

n.NA: number of missing values in x

d.mode: (weighted) mode (most common value(s))
}
