% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-baseline.R, R/fncs-desc.R
\name{bnry.count_prop.bl}
\alias{bnry.count_prop.bl}
\alias{desc-bnry}
\alias{bnry.count_prop}
\title{'bnry' describers}
\usage{
bnry.count_prop.bl(x, weight = NULL, ...)

bnry.count_prop(x, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}
}
\description{
functions to describe bnry variables
}
\details{
bnry.count_prop.bl: (weighted) count and proportion of
non-reference value as single string

bnry.count_prop: (weighted) count and proportion of non-reference
}
