% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-baseline.R, R/fncs-desc.R
\name{lcat.bl}
\alias{lcat.bl}
\alias{desc-lcat}
\title{'lcat' describers}
\usage{
lcat.bl(x, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}
}
\description{
functions to describe lcat variables
}
\details{
lcat.bl: number of unique non-missing values in x
}
