% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc.R
\name{desc-surv}
\alias{desc-surv}
\alias{eventrate}
\title{'surv' describers}
\usage{
eventrate(time, event, weight = NULL, time.unit = NULL, ...)
}
\arguments{
\item{time}{time component of time-to-event variable}

\item{event}{event component of time-to-event variable}

\item{weight}{case weight}

\item{time.unit}{numeric; if provided 'time' will be divided by this number}

\item{...}{arguments passed}
}
\description{
functions to describe surv variables
}
\details{
eventrate: (weighted) time, events and rate thereof
}
