% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dmissing}
\alias{dmissing}
\title{Missing stats}
\usage{
dmissing(data, guide = NULL, gtab = NULL)
}
\arguments{
\item{data}{data.frame}

\item{guide}{passed to \code{dable}}

\item{gtab}{passed to \code{dable}}
}
\description{
Count and percentage of missing
}
