% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy_defaults.R
\name{dpset}
\alias{dpset}
\alias{dpset_list}
\alias{dpget}
\alias{dpget_all}
\alias{dp_restore_defaults}
\alias{add2bnry.list}
\title{set package parameter}
\usage{
dpset(param, value = NULL)

dpset_list(l)

dpget(param, verbose = FALSE)

dpget_all()

dp_restore_defaults()

add2bnry.list(value)
}
\arguments{
\item{param}{character; name of parameter}

\item{value}{the value to be set}

\item{l}{list of values, the names of which correspond to parameter names}

\item{verbose}{logical; get(possibly) helpful messages?}
}
\description{
set a dable package parameter
}
\section{Functions}{
\itemize{
\item \code{dpset_list()}: dpset_list: set all paramaters in a list

\item \code{dpget()}: dpget: get package parameter

\item \code{dpget_all()}: dpget_all: get all package parameters

\item \code{dp_restore_defaults()}: dp_restore_defaults: restores the dable package parameter defaults

\item \code{add2bnry.list()}: add2bnry.list: add a value to the 'bnry.list'

}}
