% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dstd}
\alias{dstd}
\title{Standardized differences}
\usage{
dstd(data, gtab, comp, ...)
}
\arguments{
\item{data}{data.frame}

\item{gtab}{grouping table}

\item{comp}{argument for 'comp' entry of list given as 'part' argument in
\code{dable}. This can be given as 'across' (all groups compared to
first group) or 'adjacent' (all groups compared to prior group). If
given as TRUE it defaults to 'across'. Note; you can also specify a list
of pairs for the indexes to be compared; e.g. if \code{gtab} defines 4
groups \code{comp = list(c(2,1),c(4,3))} will have the second group
compared to the first, and the fourth group compared to the third.}

\item{...}{arguments passed to \code{dable}}
}
\description{
Calculate the standardized differences between groups defined by gtab.
}
