% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable2flextable.R
\name{flexdable}
\alias{flexdable}
\title{turn dable into a flextable}
\usage{
flexdable(
  dt,
  format = TRUE,
  size = "rows",
  kill = "term",
  grey = "term",
  row.group = NULL,
  fontsize = c(11, 11, 9),
  insert.bottom = TRUE
)
}
\arguments{
\item{dt}{object created by dable}

\item{format}{logical; format the variables in dt?}

\item{size}{character; what entity is presentet in row header 'rows',
'units' or 'weight'}

\item{kill}{character vector; columns to remove for presentation}

\item{grey}{character; which rows to make grey-ish in presentation}

\item{row.group}{logical or NULL; use the grouping in guide? If NULL this
will be algorithmically decided}

\item{fontsize}{numerical vector of length 3 for fontsize of header, body
and footnotes}

\item{insert.bottom}{character or logical; text to be placed under the
table. If TRUE, attr2text will be used to create this text}
}
\description{
A flextable can be used to create several types of output
}
