% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{gmatch}
\alias{gmatch}
\title{grepl matching}
\usage{
gmatch(x, table, right = TRUE, nomatch = NA_integer_, ...)
}
\arguments{
\item{x}{vector; the values to be matches}

\item{table}{vactor; the values to be matched against}

\item{right}{logical; if \code{TRUE} behaviour is like \code{match}, but if
\code{FALSE} then it is the \code{table} elements that are used as the
\code{pattern} argument in \code{grepl}, i.e. this reverses the
direction of finding matches. Note however, that the function always
returns the indices of the \code{table} argument. See example for
possible use cases.}

\item{nomatch}{the value to be returned in the case when no match is
found. Note that it is coerced to integer}

\item{...}{arguments passed to \code{grep}}
}
\value{
An integer vector of the same length as \code{x}, giving the indices
of the elements in \code{table} which matched, or \code{nomatch}.
}
\description{
\code{gmatch} returns a vector of the positions of (first)
matches of its first argument in its second.
}
\details{
Similar to \code{match} but uses \code{grep} to determine matches.
}
\examples{
## if matches are exact (and there is no risk of matching confusion) then
## match and gmatch behave the same (even if right = FALSE)
x <- LETTERS[c(2,2,4,6,1,1,3)]
y <- template <- LETTERS[5:1]
match(x, y)
gmatch(x, table = y, right = TRUE)
gmatch(x, table = y, right = FALSE)
## finding matches with gmatch could be useful if e.g. there might be a case
## mismatch
x <- c("bar", "foo", "bar")
y <- c("Foo", "Bar")
match(x, y) ## this "fails"
match(x, tolower(y)) ## this works
gmatch(x, table = y, ignore.case = TRUE)
## Sometimes there is a risk of "bad" matches
x <- c("bar", "foo", "bar")
y <- c("Foo (not Bar)", "Bar")
gmatch(x, table = y, ignore.case = TRUE) ## probably not what is wanted
## to remedie this, one need to wrap the input in a regular expression
## complex enough to handle the situation, here e.g. to make sure that only
## the inital part of the string is matched
gmatch(sprintf("^\%s", x), table = y, ignore.case = TRUE)
## sometimes it is useful to be able to reverse the direction of the match
## (even though some of these could possibly be handled with a more
## complex regular expression), e.g.
x <- c("bar a", "foo x", "foo y", "bar b", "foo z")
y <- c("foo", "bar")
gmatch(x, table = y, right = FALSE)
}
