% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fncs-gtab.R
\name{gtab-fncs}
\alias{gtab-fncs}
\alias{gtab_maker}
\title{group table (gtab) functions}
\usage{
gtab_maker(
  term,
  data,
  all = FALSE,
  all.first = FALSE,
  all.nm = NULL,
  rev = FALSE
)
}
\arguments{
\item{term}{character; name of variable in data}

\item{data}{a data.frame}

\item{all}{logical; use all rows as one of the groups}

\item{all.first}{logical; put 'all' first}

\item{all.nm}{character; name of 'all' column}

\item{rev}{logical; reverse order of the levels of 'term}
}
\description{
group table (gtab) functions
}
\details{
gtab_maker: create grouping table from a term in data
}
