% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy_defaults.R
\name{parameter-fncs}
\alias{parameter-fncs}
\alias{dpset}
\alias{dpset_defaults}
\alias{dpget}
\alias{dpget_all}
\alias{add2bnry.list}
\title{dable parameter functions}
\usage{
dpset(param, value = NULL)

dpset_defaults(overwrite = TRUE, style = NA_character_)

dpget(param, verbose = FALSE, sub = TRUE)

dpget_all()

add2bnry.list(bl)
}
\arguments{
\item{param}{character; name of parameter OR a named list of parameter values}

\item{value}{the value to be set}

\item{overwrite}{logical; overwrite parameter values already set?}

\item{verbose}{logical; get (possibly) helpful messages?}

\item{sub}{logical; search for possible replacement values?}

\item{bl}{list of binary values}
}
\description{
functions to set and retrieve dable parameters
}
\details{
set a dable package parameter

set default parameter values

get package parameter

get all package parameters

add a value to the 'bnry.list'
}
