% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{peek at an object}
\usage{
peek(x, right = FALSE, ...)
}
\arguments{
\item{x}{object to take a peek at}

\item{right}{logical, passed to print}

\item{...}{passed to print}
}
\description{
a printing method for interactive use only - do not use in scripts as it is
likely to be unstable
}
