% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dguide.R
\name{subset.dguide}
\alias{subset.dguide}
\title{subset a dguide object}
\usage{
\method{subset}{dguide}(x, ...)
}
\arguments{
\item{x}{dguide object}

\item{...}{passed to subset.data.frame}
}
\description{
subset a dguide object
}
