% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-test.R
\name{test-all}
\alias{test-all}
\alias{noTest}
\alias{noTest.bl}
\title{general testers}
\usage{
noTest(x, g, ...)

noTest.bl(x, g, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{...}{arguments passed}
}
\description{
functions to test any variables

noTest: no testing

noTest.bl: no testing (for baseline tables)
}
