% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-test.R
\name{test-bnry}
\alias{test-bnry}
\alias{bnry.chisq}
\alias{bnry.chisq.bl}
\title{'bnry' testers}
\usage{
bnry.chisq(x, g, ...)

bnry.chisq.bl(x, g, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}
}
\description{
functions to test bnry variables

bnry.chisq: chisquare-test for binary variables

bnry.chisq.bl: chisquare-test for binary variables (baseline)
}
