% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-test.R
\name{test-catg}
\alias{test-catg}
\alias{catg.chisq}
\alias{catg.chisq.bl}
\title{'catg' testers}
\usage{
catg.chisq(x, g, ...)

catg.chisq.bl(x, g, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{...}{arguments passed}
}
\description{
functions to test catg variables

catg.chisq:  chisquare-test for categorical variables

catg.chisq.bl: chisquare-test for categorical variables (baseline)
}
