% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-test.R
\name{test-date}
\alias{test-date}
\alias{date.nonparam}
\alias{date.nonparam.bl}
\title{'date' testers}
\usage{
date.nonparam(x, g, ...)

date.nonparam.bl(x, g, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{...}{arguments passed}
}
\description{
functions to test date variables

date.nonparam: non-parametric test of dates treated as integers

date.nonparam.bl: non-parametric test of dates treated as
integers (for baseline tables)
}
