% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-test.R
\name{test-real}
\alias{test-real}
\alias{param}
\alias{param.bl}
\alias{nonparam}
\alias{nonparam.bl}
\title{'real' testers}
\usage{
param(x, g, ...)

param.bl(x, g, ...)

nonparam(x, g, ...)

nonparam.bl(x, g, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}
}
\description{
functions to test real variables

param: parametric test via linear model

param.bl: parametric test via linear model (for baseline tables)

nonparam: non-parametric test

nonparam.bl: non-parametric test (for baseline tables)
}
