% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-test.R
\name{test-surv}
\alias{test-surv}
\alias{logrank}
\alias{logrank.bl}
\title{'surv' testers}
\usage{
logrank(time, event, g, ...)

logrank.bl(time, event, g, ...)
}
\arguments{
\item{time}{time component of time-to-event variable}

\item{event}{event component of time-to-event variable}

\item{g}{grouping variable}
}
\description{
functions to test surv variables

logrank: logrank test of survival curves

logrank.bl: logrank test of survival curves (for baseline tables)
}
