% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-data.R
\name{test_data}
\alias{test_data}
\alias{test_vtab}
\alias{test_stab}
\title{test data}
\usage{
test_data(n = 1000, seed = 1)

test_vtab(include.tte = FALSE)

test_stab()
}
\arguments{
\item{n}{integer; number of rows (at least 10)}

\item{seed}{numeric; seed for random number generator}

\item{include.tte}{logical; include time-to-event variables? These are
perhaps better kept track of in a 'stab'}
}
\value{
A data.frame
}
\description{
Generate a test data set
}
\details{
Generate a test data set with a number of different types of variables.
}
\section{Functions}{
\itemize{
\item \code{test_vtab()}: test_vtab: generate an 'vtab' that accompanies the data set
generated by test_data. It contains information (label, group) for (almost)
each term.

\item \code{test_stab()}: test_stab: generate an 'stab' that accompanies the data set
generated by test_data. It contains information (label, time, event) for the
relevant variables

}}
