% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_list.R
\name{text_list}
\alias{text_list}
\title{collapse vector to a text list}
\usage{
text_list(x, and = TRUE, set = FALSE, collapse = ", ")
}
\arguments{
\item{x}{vector; the text elements you want to collapse to a list}

\item{and}{logical; should there be and 'and' between the penultimate and
last entry of x when collapsed into a list?}

\item{set}{logical; wrap result in '{}'?}

\item{collapse}{character, passed to paste}
}
\value{
character (length 1)
}
\description{
This function will collapse a vector c("A", "B", "C") to a text list,
e.g. "A, B, C" or "A, B, and C" depending on the arguments.
}
\examples{
x <- LETTERS[1:3]
text_list(x, and = TRUE, set = TRUE)
text_list(x, and = TRUE, set = FALSE)
text_list(x, and = FALSE, set = TRUE)
text_list(x, and = FALSE, set = FALSE)
}
