% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc-BL-2.R
\name{baseline-standard-2}
\alias{baseline-standard-2}
\alias{real.bl2}
\alias{catg.bl2}
\alias{lcat.bl2}
\alias{date.bl2}
\alias{surv.bl2}
\title{standard baseline functions 2}
\usage{
real.bl2(x, weight = NULL, ...)

catg.bl2(x, weight = NULL, ...)

lcat.bl2(x, ...)

date.bl2(x, ...)

surv.bl2(time, event, weight = NULL, time.unit = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}

\item{...}{arguments passed along}
}
\description{
Description functions that work together
}
\details{
stats.bl2: (weighted) various stats

catg.bl2: (weighted) count and percent

lcat.bl2: unique values

date.bl2: min, max

surv.bl2: Event, rate
}
