% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-baseline.R
\name{baseline-standard}
\alias{baseline-standard}
\alias{mean_sd.bl}
\alias{mediqr.bl}
\alias{bnry.count_prop.bl}
\alias{catg.count_prop.bl}
\alias{lcat.bl}
\alias{date.bl}
\alias{eventrate.bl}
\title{standard baseline functions}
\usage{
mean_sd.bl(x, weight = NULL, ...)

mediqr.bl(x, weight = NULL, ...)

bnry.count_prop.bl(x, weight = NULL, ...)

catg.count_prop.bl(x, weight = NULL, ...)

lcat.bl(x, ...)

date.bl(x, ...)

eventrate.bl(time, event, weight = NULL, time.unit = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}

\item{...}{arguments passed along}
}
\description{
Description functions that work together
}
\details{
mean_sd.bl: (weighted) mean and standard deviation as single string

midiqr.bl: (weighted) median and interquartile range as single string

bnry.count_prop.bl: (weighted) count and proportion of
non-reference value as single string

catg.count_prop.bl: (weighted) count and proportion of
categorical levels as single string

lcat.bl: number of unique non-missing values in x

date.bl: min/max

eventrate.bl: events; rate
}
