% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc-BL-1.R
\name{baseline-standard}
\alias{baseline-standard}
\alias{real.bl0}
\alias{real.bl1}
\alias{bnry.bl0}
\alias{catg.bl0}
\alias{lcat.bl0}
\alias{date.bl0}
\alias{surv.bl0}
\title{default baseline functions (0-1)}
\usage{
real.bl0(x, weight = NULL, ...)

real.bl1(x, weight = NULL, ...)

bnry.bl0(x, weight = NULL, ...)

catg.bl0(x, weight = NULL, ...)

lcat.bl0(x, ...)

date.bl0(x, ...)

surv.bl0(time, event, weight = NULL, time.unit = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}

\item{...}{arguments passed along}

\item{time}{time component of time-to-event variable}

\item{event}{event component of time-to-event variable}

\item{time.unit}{unit of time to present in}
}
\description{
Description functions that work together. Function sets 0 and 1 are
identical except for type real; in set 0 description is by median (Q1-Q3)
and in set 1 by mean (standard deviation).
}
\details{
real.bl0: (weighted) median and interquartile range

real.bl1: (weighted) mean and standard deviation

bnry.bl0/1: (weighted) count and percent of
non-reference value as single string

catg.bl0/1: (weighted) count and proportion of
categorical levels as single string

lcat.bl0/1: number of unique non-missing values in x

date.bl0/1: min/max

surv.bl0/1: events; rate
}
