% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc-BL-1.R
\name{bl_theme}
\alias{bl_theme}
\title{Set baseline functions}
\usage{
bl_theme(desc = "default", comp = NULL, test = NULL)
}
\arguments{
\item{desc}{the 'theme' for descriptives}

\item{comp}{the 'theme' for comparisons (not yet implemented)}

\item{test}{the 'theme' for testers (not yet implemented)}
}
\description{
Quickly set baseline functions to some predefined set of functions.
}
