% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-bnry}
\alias{comp-bnry}
\alias{bnry.std}
\title{'bnry' comparers}
\usage{
bnry.std(x, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to compare bnry variables

bnry.std: (weighted) standardized difference for 'bnry'
}
