% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-catg}
\alias{comp-catg}
\alias{catg.std}
\alias{catg.diff}
\title{'catg' comparers}
\usage{
catg.std(x, g, weight = NULL, catg.full.length = FALSE, ...)

catg.diff(x, g, weight = NULL, catg.full.length = FALSE, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{catg.full.length}{logical; return dimension equal to the input (is
this EVER wanted?)}

\item{...}{arguments passed}
}
\description{
functions to compare catg variables

catg.std: (weighted) standardized difference for 'catg'

catg.diff: (weighted) difference in proportion
}
