% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp-STDIFF.R, R/fncs-comp.R
\name{catg.std}
\alias{catg.std}
\alias{comp-catg}
\title{'catg' comparers}
\usage{
catg.std(x, g, weight = NULL, catg.full.length = FALSE, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}
}
\description{
functions to compare catg variables
}
\details{
catg.std: (weighted) standardized difference for 'catg'
}
