% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-date}
\alias{comp-date}
\alias{date.std}
\title{'date' comparers}
\usage{
date.std(x, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}
}
\description{
functions to compare date variables
}
\details{
date.std: (weighted) standardized difference for 'date' ( = real.std
applied to x interpreted as an integer)
}
