% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-date}
\alias{comp-date}
\alias{date.std}
\alias{date.diff}
\title{'date' comparers}
\usage{
date.std(x, g, weight = NULL, ...)

date.diff(x, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to compare date variables

date.std: (weighted) standardized difference for 'date' ( = real.std
applied to x interpreted as an integer)

date.diff: (weighted) difference for 'date' ( = real.diff
applied to x interpreted as an integer)
}
