% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp-STDIFF.R, R/fncs-comp.R
\name{real.std}
\alias{real.std}
\alias{comp-real}
\title{'real' comparers}
\usage{
real.std(x, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}
}
\description{
functions to compare real variables
}
\details{
real.std: (weighted) standardized difference
}
