% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-real}
\alias{comp-real}
\alias{real.std}
\alias{mean_diff}
\title{'real' comparers}
\usage{
real.std(x, g, weight = NULL, ...)

mean_diff(x, g, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to compare real variables

real.std: (weighted) standardized difference for 'real'

mean_diff: (weighted) difference in mean value
}
