% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-surv}
\alias{comp-surv}
\alias{surv.std}
\alias{surv.diff}
\alias{rate_ratio}
\alias{rate_ratio.bl}
\alias{hazard_ratio}
\alias{hazard_ratio.bl}
\alias{hr_ci}
\alias{hr_ci.bl}
\title{'surv' comparers}
\usage{
surv.std(time, event, g, weight = NULL, ...)

surv.diff(time, event, g, weight = NULL, ...)

rate_ratio(time, event, g, weight = NULL, ...)

rate_ratio.bl(time, event, g, weight = NULL, ...)

hazard_ratio(time, event, g, weight = NULL, ...)

hazard_ratio.bl(time, event, g, weight = NULL, ...)

hr_ci(time, event, g, weight = NULL, ...)

hr_ci.bl(time, event, g, weight = NULL, ...)
}
\arguments{
\item{time}{time component of time-to-event variable}

\item{event}{event component of time-to-event variable}

\item{g}{grouping variable}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to compare surv variables

surv.std: (weighted) standardized difference for 'surv'

surv.diff: (weighted) rate difference for 'surv'

rate_ratio.bl: ratio of the (weighted) rates (baseline)

hazard_ratio.bl: (weighted) hazard ratio (baseline)

hr_ci.bl: (weighted) hazard ratio and confidence interval (baseline)
}
\details{
rate_ratio: ratio of the (weighted) rates

hazard_ratio: (weighted) hazard ratio

hr_ci:  (weighted) hazard ratio and confidence interval
}
