% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-comp.R
\name{comp-surv}
\alias{comp-surv}
\alias{surv.std}
\alias{rate_ratio}
\title{'surv' comparers}
\usage{
surv.std(time, event, g, weight = NULL, ...)

rate_ratio(time, event, g, weight = NULL, ...)
}
\arguments{
\item{time}{time component of time-to-event variable}

\item{event}{event component of time-to-event variable}

\item{g}{grouping variable}

\item{weight}{case weight}
}
\description{
functions to compare surv variables

surv.std: (weighted) standardized difference for
'surv'
}
\details{
rate_ratio: ratio of the (weighted) rates
}
