% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{dable}
\alias{dable}
\alias{dreal}
\alias{dsurv}
\alias{ddate}
\alias{dcatg}
\alias{dbnry}
\alias{dlcat}
\title{Data description}
\usage{
dable(
  data,
  type = "baseline",
  bl.rm = NULL,
  guide = NULL,
  gtab = NULL,
  part = list(desc = TRUE, comp = NA, test = NA),
  fnc = list(desc = NULL, comp = NULL, test = NULL),
  weight = NULL,
  ...
)

dreal(data, ...)

dsurv(data, ...)

ddate(data, ...)

dcatg(data, ..., bnry = FALSE, lcat = FALSE)

dbnry(data, ...)

dlcat(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{type}{character; either a data type or 'baseline' for all types}

\item{bl.rm}{character vector; types to remove from baseline table}

\item{guide}{object created with \code{dguide}}

\item{gtab}{a grouping table which is a named list of logical vectors equal
in length to the numbers of rows of data. It can also be created by
pointing to a term in data.}

\item{part}{list that specifies which 'parts' of the table you want. The
parts are for 'description', 'comparison' and 'test'. There is more
funcionality hidden here, but that must be explained in a vignette.}

\item{fnc}{list of functions for the different parts. Currently only
supported for a non-baseline type.}

\item{weight}{a vector of weights, or the name of a term in data that
provides these}

\item{...}{arguments passed to functions that describes, compares and tests}

\item{bnry}{logical; for catg-shorthand, also include bnry type?}

\item{lcat}{logical; for catg-shorthand, also include lcat type?}
}
\description{
This function generates a data description.
}
\section{Functions}{
\itemize{
\item \code{dreal()}: dreal is a shorthand for \code{dable(type = 'real')}

\item \code{dsurv()}: dsurv is a shorthand for \code{dable(type = 'surv')}

\item \code{ddate()}: ddate is a shorthand for \code{dable(type = 'date')}

\item \code{dcatg()}: dcatg is a shorthand for \code{dable(type = 'catg')}

\item \code{dbnry()}: dbnry is a shorthand for \code{dable(type = 'bnry')}

\item \code{dlcat()}: dlcat is a shorthand for \code{dable(type = 'lcat')}

}}
