% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipher.R
\name{decipher}
\alias{decipher}
\title{replace selected values}
\usage{
decipher(x, key, flexible = TRUE, within = FALSE, within_word = FALSE)
}
\arguments{
\item{x}{Character (or factor) vector}

\item{key}{Named character vector; where (typically) the names corresponds
to values appearing in x, and the values are those we want as
replacements. However, if \code{flexible = TRUE}, then a check is
performed to see if it makes more sense to invert the key. It is also
possible to supply a data.frame with columns 'term' and 'label', if so a
key is created from this information.}

\item{flexible}{Logical; if TRUE the key can be inverted}

\item{within}{Logical; if TRUE occurences of the keyed values within a
string can be replaced.}

\item{within_word}{Logical; if 'within' is TRUE, one can use this parameter
to choose if replacement can occur within words or not. E.g. if TRUE
then replacing 'break' with 'pause' will turn 'breakfast' into
'pausfast'. I imagine this is not typically what is wanted. If FALSE,
replacement occurences must be encapsulated by beginning of line, space
characters, punctuation characters or end of line, so that replacement
does not take place within words.}
}
\value{
vector of same length and class as s (character or factor)
}
\description{
Change values that corresponds to those of a key.
}
\examples{
x <- c("foo", "foo!", "A foo", "A foo and two bar",
       "football", "barely", "bar")
key <- c(foo = "yay", bar = "bohnanza")
data.frame(
  x = x,
  d0 = key[x], ## useful only if all values of s exists in the key
  d1 = decipher(x, key),
  d2 = decipher(x, key, within = TRUE)
)
x <- factor(c(1,3,2), levels = 1:3,
            labels = c("foo", "bar", "baz"))
key <- c(A = "foo", C = "baz")
str(data.frame(x = x, d = decipher(x, key)))
}
