% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc.R
\name{desc-bnry}
\alias{desc-bnry}
\alias{bnry.count_prop}
\title{'bnry' describers}
\usage{
bnry.count_prop(x, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}

\item{...}{arguments passed}
}
\description{
functions to describe bnry variables
}
\details{
bnry.count_prop: (weighted) count and proportion of non-reference
}
