% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc.R
\name{desc-catg}
\alias{desc-catg}
\alias{catg.count_prop}
\title{'catg' describers}
\usage{
catg.count_prop(x, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}
}
\description{
functions to describe catg variables
}
\details{
catg.count_prop: (weighted) count and proportion of categorical
levels
}
