% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-baseline.R, R/fncs-desc.R
\name{date.bl}
\alias{date.bl}
\alias{desc-date}
\title{'date' describers}
\usage{
date.bl(x, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}
}
\description{
functions to describe date variables
}
\details{
date.bl: min/max
}
