% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-desc.R
\name{desc-real}
\alias{desc-real}
\alias{d.min}
\alias{d.max}
\alias{min_max}
\alias{d.mean}
\alias{d.sd}
\alias{mean_sd}
\title{'real' describers}
\usage{
d.min(x, ...)

d.max(x, ...)

min_max(x, ...)

d.mean(x, weight = NULL, ...)

d.sd(x, weight = NULL, ...)

mean_sd(x, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}
}
\description{
functions to describe real variables
}
\details{
d.min: minimum

d.max: maximum

min_max: minimum and maximum

d.mean: (weighted) mean value

d.sd: (weighted) standard deviation

mean_sd: (weighted) mean and standard deviation
}
