% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs-baseline.R, R/fncs-desc.R
\name{mean_sd.bl}
\alias{mean_sd.bl}
\alias{desc-real}
\alias{d.min}
\alias{d.max}
\alias{min_max}
\alias{d.mean}
\alias{d.sd}
\alias{mean_sd}
\alias{d.quantile}
\alias{quartile}
\title{'real' describers}
\usage{
mean_sd.bl(x, weight = NULL, ...)

d.min(x, ...)

d.max(x, ...)

min_max(x, ...)

d.mean(x, weight = NULL, ...)

d.sd(x, weight = NULL, ...)

mean_sd(x, weight = NULL, ...)

d.quantile(x, weight = NULL, probs = c(0.25, 0.5, 0.75), ...)

quartile(x, weight = NULL, ...)
}
\arguments{
\item{x}{input vector}

\item{weight}{case weight}
}
\description{
functions to describe real variables
}
\details{
mean_sd.bl: (weighted) mean and standard deviation as single string

d.min: minimum

d.max: maximum

min_max: minimum and maximum

d.mean: (weighted) mean value

d.sd: (weighted) standard deviation

mean_sd: (weighted) mean and standard deviation

d.quantile: (weighted) quantiles (wrapper for Hmisc::wtd.quantile)

quartiles: (weighted) quartiles
}
